+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [3.7.0] - 2024-04-04
+
 ### Added
 
 - Add managed object storage name to `show` and `list` command outputs.
@@ -370,7 +372,8 @@ Initial public beta release :tada:
 ### Added
 - Current feature set added! First internal release
 
-[Unreleased]: https://github.com/UpCloudLtd/upcloud-cli/compare/v3.6.0...HEAD
+[Unreleased]: https://github.com/UpCloudLtd/upcloud-cli/compare/v3.7.0...HEAD
+[3.7.0]: https://github.com/UpCloudLtd/upcloud-cli/compare/v3.6.0...v3.7.0
 [3.6.0]: https://github.com/UpCloudLtd/upcloud-cli/compare/v3.5.0...v3.6.0
 [3.5.0]: https://github.com/UpCloudLtd/upcloud-cli/compare/v3.4.0...v3.5.0
 [3.4.0]: https://github.com/UpCloudLtd/upcloud-cli/compare/v3.3.0...v3.4.0